Adds orders api/events/ui files
Adds api, events, and ui files for front end.  Adds event listener to
product-list table generated by handlebars to listen for Add to Cart
button submit.  Fires onOrderUpdate on the Add to Cart submit to send
a successful patch request to update an order with the selected product.

The order id is currently hardcoded and needs to be updated once the
order creation workflow has been created.